@@ -33,7 +33,9 @@
     "@babel/preset-typescript": "^7.22.5",
     "@commitlint/cli": "^13.2.1",
     "@commitlint/config-conventional": "^13.2.0",
+    "@reduxjs/toolkit": "^1.9.5",
     "@types/react-dom": "^18.2.6",
+    "@types/webpack-env": "^1.18.1",
     "@typescript-eslint/parser": "^6.1.0",
     "add-asset-html-webpack-plugin": "^3.1.3",
     "antd": "^4.16.13",
@@ -67,8 +69,8 @@
     "postcss-loader": "4.2.0",
     "postcss-scss": "^4.0.6",
     "prettier": "^2.6.2",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-redux": "^7.2.0",
     "react-router-dom": "^6.9.0",
     "redux": "^4.0.5",
@@ -86,10 +88,10 @@
     "ts-loader": "~8.2.0",
     "typescript": "^5.1.3",
     "url-loader": "^4.1.0",
-    "webpack": "^4.43.0",
+    "webpack": "^5.88.2",
     "webpack-bundle-analyzer": "^4.6.1",
-    "webpack-cli": "^3.3.11",
-    "webpack-dev-server": "^3.10.3",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^4.15.1",
     "webpack-merge": "^4.2.2"
   },
   "dependencies": {
